#14: Refactor gruntfile.js
