+++ grunt/configs.js
@@ -1,0 +1,5 @@
+const eslint = require('./eslint');
+
+module.exports = {
+    eslint
+};
