+++ package.json
@@ -24,8 +24,6 @@
     "eslint-formatter-pretty": "^0.2.2",
     "grunt": "^1.0.1",
     "grunt-eslint": "KK578/grunt-eslint",
-    "jit-grunt": "^0.10.0",
-    "load-grunt-config": "^0.19.2",
     "time-grunt": "^1.4.0"
   }
 }
